подправила кнопку выключения формы
