+++ js/main.js
@@ -1,6 +1,6 @@
 import '../js/miniatures.js';
 import '../js/data.js';
-import '../js/form.js'
+import '../js/form.js';
 import {openPicture} from './big-image.js';
 
 openPicture();
